app/requirements.txt
Flask
fluent-logger
